Improve existing plugs for authentication and unauthenticated routes
