--- lib/crawlexir_web/plugs/auth.ex
@@ -1,30 +1,0 @@
-defmodule CrawlexirWeb.Plugs.Auth do
-  import Plug.Conn
-  import Phoenix.Controller
-
-  alias Crawlexir.Auth
-
-  def init(opts), do: opts
-
-  @doc """
-  Redirect non-authenticated users to the sign in page
-  """
-  def call(conn, _opts) do
-    case get_session(conn, :current_user_id) do
-      nil ->
-        conn
-        |> assign(:current_user, nil)
-        |> assign(:user_signed_in?, false)
-        |> put_flash(:info, "Login into your account.")
-        |> redirect(to: CrawlexirWeb.Router.Helpers.session_path(conn, :new))
-        |> halt()
-
-      user_id ->
-        current_user = Auth.get_user!(user_id)
-
-        conn
-        |> assign(:current_user, current_user)
-        |> assign(:user_signed_in?, true)
-    end
-  end
-end
